Cargo.toml: upgrade to version 0.0.3
We'll try this thing that some projects do (e.g. rustc): odd version
numbers will be for dev version, even numbers for stables.

Well, we're not stable yet, so we're using the third component of
version numbers. And odd numbers will be for the dev version on GitHub,
while even version will be used for crates pushed to crates.io so that
their number only identifies a single version of the crate.