src/asm_parser.rs: remove public exposure of the module in root file
There seems to be no particular reason to give access to this module to
the user, since all the assembler API is held by the `assembler` module.

Rust-doc comments were left in place. I added in a previous commit, and
removed here, an example test for the `parse()` function, just in case
we ever make this module public again (nice to have examples for all
exposed functions).

All unit tests from `tests/asm_parser.rs` had to been moved to the
`tests` module of the `src/asm_parser.rs` file.

Fixes: #13