Cargo.toml: update dependencies
Update dependencies as long as it does not break the lib. In fact, there
is just one dependency that cannot be updated to its latest version:
"combine", currently at 3.6.0, with major API changes between v2 and v3.
Let's stick to v2 for now.

Thanks to @jackcmay for this update.

Fixes #35.