src/lib.rs: minor cleanup in documentation
Following the addition of customized verifier support, bring minor
modifications to the documentation of the library. In particular, remove
comments about set_verifier() that would panic (it does not, it returns
an error on verification failure).

Also remove spurious blank lines at the end of a test file (I was just
asking for an end-of-line char at the end of the last line ;) ).